+++ apps/admin/src/layout/app-sidebar/NavGhostPro.tsx
@@ -1,0 +1,31 @@
+import React from "react"
+
+import {
+    LucideIcon,
+    SidebarGroup,
+    SidebarGroupContent,
+    SidebarMenu,
+    SidebarMenuButton,
+    SidebarMenuItem
+} from "@tryghost/shade"
+
+function NavGhostPro({ ...props }: React.ComponentProps<typeof SidebarGroup>) {
+    return (
+        <SidebarGroup {...props}>
+            <SidebarGroupContent>
+                <SidebarMenu>
+                    <SidebarMenuItem>
+                        <SidebarMenuButton asChild>
+                            <a href="#/billing">
+                                <LucideIcon.CreditCard />
+                                <span>Ghost(Pro)</span>
+                            </a>
+                        </SidebarMenuButton>
+                    </SidebarMenuItem>
+                </SidebarMenu>
+            </SidebarGroupContent>
+        </SidebarGroup>
+    );
+}
+
+export default NavGhostPro;
